move icons to where they are rendered, rename UppyIcon to uppy-u-icon
