+++ packages/@uppy/core/src/_common.scss
@@ -1,5 +1,5 @@
 /**
-* General Uppy styles that apply to everything inside the .Uppy container
+* General Uppy styles that apply to everything inside the .uppy-Root container
 */
 
 .uppy-Root {
@@ -19,16 +19,6 @@
 
 .uppy-Root [hidden] {
   display: none;
-}
-
-// https://blog.prototypr.io/align-svg-icons-to-text-and-say-goodbye-to-font-icons-d44b3d7b26b4
-
-.UppyIcon {
-  max-width: 100%;
-  max-height: 100%;
-  fill: currentColor; /* no !important */
-  display: inline-block;
-  overflow: hidden;
 }
 
 // Utilities
@@ -120,6 +110,17 @@
     border-color: $gray-700;
     box-shadow: none;
   }
+}
+
+// Icon
+
+// https://blog.prototypr.io/align-svg-icons-to-text-and-say-goodbye-to-font-icons-d44b3d7b26b4
+.uppy-c-icon {
+  max-width: 100%;
+  max-height: 100%;
+  fill: currentColor;
+  display: inline-block;
+  overflow: hidden;
 }
 
 // Buttons
