+++ packages/@uppy/dashboard/src/components/FileItem/FileProgress/index.scss
@@ -34,7 +34,7 @@
     }
   }
 
-.UppyIcon-progressCircle {
+.uppy-c-icon-progressCircle {
   width: 100%;
   height: 100%;
